feat: enhance the vouchers api (#53)
* feat: enhance the vouchers api

Basically have it list all vouchers with their fields too so an admin
can have a better look at what it has.
This is also needed to better export vouchers, for example, somebody
provisions a system and knows the MAC used as device_info and they can
thus just grab that directly (real world use case). Just relying on GUID
is not very handy as when a device is initialized and a voucher created
in the db, there's no way to link them together.

Signed-off-by: Antonio Murdaca <antoniomurdaca@gmail.com>

* fix: add sane TTL to to0

Signed-off-by: Antonio Murdaca <antoniomurdaca@gmail.com>

* fix: adapt to the new voucher fetch api

Signed-off-by: Antonio Murdaca <antoniomurdaca@gmail.com>

* fix: avoid a double Scan call

Co-authored-by: Miguel Martin <mmartinv@redhat.com>
Signed-off-by: Antonio Murdaca <antoniomurdaca@gmail.com>

---------

Signed-off-by: Antonio Murdaca <antoniomurdaca@gmail.com>
Co-authored-by: Miguel Martin <mmartinv@redhat.com>